feat: Refactor Enable And Disable Module Commands (#1918)
* feat: Refactor Enable And Disable Module Commands

* feat: Update docs

* fix: Unrelated Unit Test

* Apply suggestions from code review


* docs: Update

* Apply suggestions from code review


* docs: Update

---------
