+++ cmd/kyma/alpha/add/add.go
@@ -1,15 +1,16 @@
-package enable
+package add
 
 import (
-	"github.com/kyma-project/cli/cmd/kyma/alpha/enable/module"
+	"github.com/kyma-project/cli/cmd/kyma/alpha/add/module"
 	"github.com/kyma-project/cli/internal/cli"
 	"github.com/spf13/cobra"
 )
 
 func NewCmd(o *cli.Options) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "enable",
-		Short: "Enables a resource in the Kyma cluster.",
+		Use:     "add",
+		Aliases: []string{"enable"},
+		Short:   "Adds a resource to the Kyma cluster.",
 		Long: `Use this command to enable a resource in the Kyma cluster.
 `,
 	}
